+++ src/main/java/view/InputView.java
@@ -1,6 +1,6 @@
 package view;
 
-import domain.Name;
+import domain.participant.Name;
 
 import java.util.Arrays;
 import java.util.List;
